Ajout de la notation d'une recette a travers les commentaires
Une recette est noté au même que quelqu'un laisse un commentaire et la note moyenne de la recette est calculé en même temps.

+++ Back/databaseTest/views/comments.py
@@ -14,20 +14,34 @@ def add_comment(request, recipe_title):
             data = json.loads(request.body)
             texte = data.get("texte")
             images = data.get("images", [])
+            note = data.get("note")
 
             if not texte:
                 return JsonResponse({"error": "Veuillez laisser un commentaire."}, status=400)
+            if not note or not (1 <= note <= 5):
+                return JsonResponse({"error": "Veuillez noter la recette entre 1 et 5."}, status=400)
+
 
             try:
                 recipe = Recipe.nodes.get(titre=recipe_title)
             except Recipe.DoesNotExist:
                 return JsonResponse({"error": "Recette non trouvée."}, status=404)
 
-            commentaire = Comments(texte=texte, images=images)
+            commentaire = Comments(texte=texte, images=images, note=note)
             commentaire.save()
             commentaire.recette.connect(recipe)
 
-            return JsonResponse({"message": "Commentaire ajouté avec succès", "comment_id": commentaire.uuid})
+            # Mettre à jour la moyenne des notes
+            recipe.total_notes += 1
+            recipe.somme_notes += note
+            recipe.note = recipe.somme_notes / recipe.total_notes
+            recipe.save()
+
+            return JsonResponse({
+                "message": "Commentaire ajouté avec succès",
+                "note_moyenne": recipe.note,
+                "comment_id": commentaire.uuid
+                })
         
         except json.JSONDecodeError:
             return JsonResponse({"error": "Données invalides."}, status=400)
@@ -40,6 +54,7 @@ def get_comment(request, comment_id):
             commentaire = Comments.nodes.get(uuid=comment_id)
             data = {
                 "texte": commentaire.texte,
+                "note": comment.note,
                 "images": commentaire.images
             }
             return JsonResponse({"commentaire": data})
@@ -57,6 +72,7 @@ def get_all_comments(request, recipe_title):
             data = [
                 {
                     "texte": commentaire.texte,
+                    "note": commentaire.note,
                     "images": commentaire.images
                 }
                 for commentaire in commentaires
@@ -74,6 +90,7 @@ def update_comment(request, comment_id):
             commentaire = Comments.nodes.get(uuid=comment_id)
             
             commentaire.texte = data.get("texte", commentaire.texte)
+            commentaire.note = data.get("note", commentaire.note)
             commentaire.images = data.get("images", commentaire.images)
             commentaire.save()
             
@@ -87,11 +104,25 @@ def update_comment(request, comment_id):
 
 # Suppression Catégorie
 @csrf_exempt
-def delete_comment(request, comment_id):
+def delete_comment(request, recette_title, comment_id):
     if request.method == "DELETE":
         try:
+            recipe = Recette.nodes.get(titre=recette_title)
             commentaire = Comments.nodes.get(uuid=comment_id)
+
+            if not recipe.commentaires.is_connected(commentaire):
+                return JsonResponse({"error": "Le commentaire n'est pas lié à cette recette."}, status=400)
+
+            recipe.commentaires.disconnect(commentaire)
+
+            # Mettre à jour la moyenne des notes
+            recipe.total_notes -= 1
+            recipe.somme_notes -= commentaire.note
+            recipe.note = recipe.somme_notes / recipe.total_notes if recipe.total_notes > 0 else 0.0
+            recipe.save()
+
             commentaire.delete()
+
             return JsonResponse({"message": "Commentaire supprimé avec succès"})
         except Comments.DoesNotExist:
             return JsonResponse({"error": "Commentaire non trouvé"}, status=404)
